X-sync After Merge And Resolve Confilicts !

+++ Application/BackgroundWorker/Common/Events/SuccessedContractAdded.cs
@@ -2,6 +2,6 @@
 {
     public class SuccessedContractAdded
     {
-        public int SuccessedContractId { get; set; }
+        public long SuccessedContractId { get; set; }
     }
 }
